Prototype for the first iteration
Added a xml with all the settings (imo only html) the admin should be able
to change and created a class to create a file with the settings.
The file will then be zipped and saved in public/profiles.
At this location it's ready to be pull from the extension.
IMO the file only contains wether HTML is enabled or not.